Added pandocShell

+++ usb/usb/editPws.sh
@@ -1,7 +1,8 @@
 #!/bin/sh
-set -e  # stop on first error
-gpg2 -d .pws.gpg > .pws
+#set -e  # stop on first error
+if gpg2 -d .pws.gpg > .pws ; then
 kak .pws
 gpg2 -c .pws
 shred .pws
 rm .pws
+fi
